python_server/requirements.txt
fastapi==0.95.1
uvicorn[standard]==0.22.0
pymongo==4.9.1
motor==3.6.0
python-multipart==0.0.5
python-decouple==3.8.0
email-validator==1.1.1
pydantic[email]==1.10.2
aiofiles==23.1.0
python-jose[cryptography]==3.3.0
openpyxl==3.1.2
pytest==7.3.1
httpx==0.24.0
pandas==2.2.3
numpy==1.24.4
PyJWT==2.7.0
